requirements.txt
# local package
-e .

# external requirements
click
Sphinx
coverage
awscli
flake8
python-dotenv>=0.5.1

#project env requirements
jupyterlab
numpy
pandas
matplotlib
plotly
python-box
pyyaml
requests
scikit-learn
scipy
seaborn
